feat: TMDB 장르 저장 및 N+1문제 해결

+++ src/main/java/com/sprint/mople/domain/content/batch/tmdb/TmdbItemDto.java
@@ -29,5 +29,5 @@ public class TmdbItemDto {
   private String posterUrl;
 
   @JsonProperty("genre_ids")
-  private List<Integer> genreIds;
+  private List<Long> genreIds;
 }
